api/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain and AI
langchain==0.1.0
langchain-openai==0.0.2
langchain-core==0.1.10
langchain-community==0.0.13
langchain-text-splitters==0.0.1
langchain-huggingface==0.0.1
openai==1.10.0

# Document processing
python-docx==1.1.0
pdfplumber==0.10.3
python-pptx==0.6.23
pandas==2.1.4
openpyxl==3.1.2

# Vector stores and embeddings
faiss-cpu==1.7.4
sentence-transformers==2.3.1

# Web search
google-api-python-client==2.111.0
beautifulsoup4==4.12.2
requests==2.31.0

# Image processing
Pillow==10.2.0

# Audio processing
SpeechRecognition==3.10.1
pydub==0.25.1

# Database
supabase==2.3.0

# Environment
python-dotenv==1.0.0

# Utilities
httpx==0.26.0